+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - If at least one task fails to build **pkger** will now return with exit code 1 [#66](https://github.com/wojciechkepka/pkger/pull/66)
 - `--quiet` flag now suppresses only container output, normal info output is still available [#67](https://github.com/wojciechkepka/pkger/pull/67)
 - `pkger build` will now display a warning with instructions on how to start a build [#68](https://github.com/wojciechkepka/pkger/pull/68)
+- Use appropriate cache dir on each OS to store images state [#71](https://github.com/wojciechkepka/pkger/pull/71)
 
 # 0.4.0
 - Add an option to sign RPMs with a GPG key [#55](https://github.com/wojciechkepka/pkger/pull/55)